Groupedit: Add back call from groupedit to widget list.
   When the groupedit item select, groupedit send a signal
"part,selected", and set to event_info ta name of part which has
been selected. The workspace hook a signal and resend it.
When in the widget list selected a part - selected item in the
groupedit. If groupedit in the separate mode, groupedit item
selected, if groupedit in the normal mode - highlight is show.