Global config for eflete.
Rename the settings module to the config. Rename the settings API to
the common config name, as config_init, config_load, config_save.
Small variable renaming in the config module. Update test accordenly
to new API.
Add profile.h it a prototype of the config profile functionality in
the eflete.