Main: adding help and usage
This feature add ability to start eflete with some parametres.
It uses ECORE_GETOPT mechanism for parsing parametres given in console to the eflete.
There are few main parametres were added: help, version, copyright,
license and the last one (-o that is the first letter of "open" or --edj)
that allow to open project (edj) file immediately with eflete

Examples:
> eflete -h
> eflete -l
> eflete --open=file.edj
> eflete -o=filename.edj
> eflete -edj=../../dir/filename.edj

@feature