UTC: fixing compile.
There are few modules and new files around (for example syntax_color in folder external)
which are required for old modules to be included.
So because of that "make check" didn't worked.
This commit fixes it by adding new folders with new modules to be included,
so tests now compiling again.

@fix