Live View Refactor: starting to refactor architecture of Live View
As the widget is starting to be too many and as there is another 
module (Live View Property) was added, there is a huge need 
in making code more readable and well organised.
Proposed this scheme of Live View.
Each widget has it's own *.c file that have next stuff in it:
- function which will create and return defined widget
- functions-callbacks and functions to set up properties of widget
- functions-callbacks for setting and unsetting texts and swallows
  (made for Live View Property)

Functions will be set into object as data, and later Live View
Property will get those functions and set as callbacks for
setting text and swallow.