Property: turn on/off ellipsis attribute
Add a toggle for turn on/off ellipsis for TEXT parts.
Insted add "-1" value to 'ellipsis' spinner, added toggle
for turn on/off a ellipsis. When a toggle is 'Off' set to
edje_edit object ellipsis value '-1', when toggle is 'On'
set a value fron spinner. By default to spinner set a 0.0.

@feature