Highlight/Workspace: changing align by moving handler within part's object area.
By pressing on Alt, workspace shows object area on a groupedit.
By pressing on Ctrl+Alt it will show middle handler in the middle of the highlight
(that follows selected part).
User can grab that handler with a mouse and move handler
(if he keeps holding Ctrl+Alt of course).

If he moves highlight like that, then the align for position of part inside of
it's relative space (so colled object area) of the whole part will be changed.

@feature