Groupedit: loading edje object with image for zoom.
If we directly set up an image then Evas_Map (for zoom purposes)
won't work properly with images.
In fact, for example, it ignores border values and params.
So in this case we need to load edje object and set image into it's swallow,
so Evas_Map can work with it properly.

@fix