Workspace: fixing Showing hidden parts on workspace (Deffect #33505)
Renaming part causes whole groupedit to be reloaded again.
That is a lot of calculation to do.
Instead there is much more easy and faster to find and rename part in groupedit,
instead of reloading and loading whole groupedit again (recreate every part
again, calculate their position, etc etc etc...)

Also fixing Deffect #35762 when if there are two parts:
"elm.swallow.icon" and "elm.swallo" when we are trying to find "elm.swallo"
it was returned "elm.swallow.icon" instead which is unacceptable.

@fix