History: fix histor_redo tests.
This fix provide changes in tests for the function history_redo accordingly to
the use UI. It mean, that now needed loaded project and initialized main window
for make any actions on the changes. It happens because in case when part is
renamed in edje_edit object it should be renamed also in Part structure, that
represented this part in widget manager module.