Sound editor: change logic of new sounds adding to the ptoject
Now the sound could be included to the project only when user pushes "Ok" button. 
But user has an ability to listen to added sounds and save or delete it later. 
If user pushes "Close" all changes made on last stage will be discarded:
the newly included sounds will be deleted from gengrid and from the list
of sounds to be included to the EDJ Ptoject on Save. 
Those operations are more quickly. This logic saves us from extra calls
to edje_edit_xxx_del functions, which have big overheads.