Property/Workspace: make 'min' and 'max' for groupedit working again.
Also, there is a problem where min is greater than max.
Then we need to update max in the way that min will be at least equal to max.

Same for a case when max is lower than min.

Also fixing 'max' value, because max can't be -1.
Default value that shows unlimit value is 0.

@fix