Live View: avoid extra setups of swallow content (Bug #37992)
Changing of state of 'Swallows' check leads to the 'changed' callback
call for each swallow checkbox in properties of Live View
but the check of current state of corresponding parts is missed.
This causes multiple contents to be created and set.
This commit contains two checks of current state of swallow part to avoid
extra operations in case is those parts are in needed states already.