Connector: update workspace to apply saved and new text styles
For example, user added lots of new styles or different data into project,
but didn't saved yet, so such data won't be apply to that.
But when he set params to that new data and click on save, it should be
applied instantly.
For example: add new style and add textblock part.
Apply that style to textblock
After clicking on save, the style should be applied automatically.

@fix