Project_manager: new signature of the pm_project_style_source_code_export
Reecently this function was used to write data to the file (code of the
current style) and also used to extract code of current style that was
used in tab "code" of the properties block. With the aim to control right
writing to the file and extract source code exactly with public library API
this function was changed.