style editor: prepare to use toolbar instead of EWE_TABS
In a far far Eflete, there were extremely uncomfortable tabs...
They were hard to support, at the same time most of things
were already living in a Toolbar (buttons for when there are too much tabs,
vertical and horizontal, tab closing, etc).
And now, on the edge of the old era, 
Great Toolbar is going to come into this worthless world.