eflete: fill the import edj tab fiels on import from command liane
Eflete it's a graphical application and do hiden import from command
line it's wrong way. So if edj file imported from the command line
Eflete show the necessary tab and fill the filds accourdingly to
used options. In this way user can correct the entered data and
can see the splash animation.

Change-Id: I6ad334423267835e7fe8043a3b257a3588c7d19e