menu: move 'scale' items to workspace ctx menu
We are forced to move those items, because in case when used the dbus
session for menu, like it did in Unity (Ubuntu), we have a problem, who
am I kidding, this a big trouble, can't set the custom widget to menu
item. So now this items in the ctx menu on workspace.

Change-Id: Icb4aad8f3dc2775ecd0ce54da65f5b4a96af0944