tabs: unpin the home tab
The home tab is unpined, and can be closed as the another tabs, means
tabs with workspace. But have one issue now - after open, create project
etc, the tab is empty. It's bug depands of two toolbars for pined and
for normal tabs. After move managers (resource manages means) to modal
window, this issue will be resolve.

Change-Id: I767594e168a783da5038a772703d322075315b9a