editor: add to func signature bool for call edje_edit func
This bool flag need for indicate editor for call edje_edit func or
nor. This is a new fuctionality need for create changes without
real changes in the edj file. And this fake changes we can move
and apply to other project. So it first step of import groups from
different files and copy past any primitives.

Change-Id: Ib15a7662420321eafae179e718c67735cf3627d6