project_navigator: fix SIGSEGV when copy group in huge project
NEVER EVER USE elm_object_item_text_get with genlist because it doesn't work when
items are unrealized. It will return NULL. That caused a lot of problems so this
function would be poisoned for now.
For getting data from unrealized items it is better to use
elm_object_item_data_get

Fixes T3335