resource_manager2: part resource delete
Need more complex check up on this code if something goes wrong or right
of course. Please recheck with cgdb if everything is alright. It seems work for
me now.
Also need more complex work up with resource deletion and addition. This is just
a prototype that works, for applying notifications and other things it would be
reworked much later.