project_close: rework save feature
Early was added the two options --export-edj and --export-edc for
Eflete. But this feature was not compleated, and was exported edj and
edc on save only if application configured with --enable-tizen.

This patch is complete this feature for both modes: normal and tizen.
Now if set --export-edj or --export-edc project data will be exported
by given path.

Change-Id: Ic00f9922e84161dd5b5565e07a53504105a8a4d7