project_manager: update pm_lock_check
Now this API check is pid in '.project.lock' file is available. If
process is running function will return that given project is locked,
if can not read pid from file that case means that have not access
permission, and it error too.

If '.project.lock' file is exist but process with pid from file not
available that means that Eflete is terminated while given project was
edit, and it project Eflete can recover on open.

In case when '.project.lock' is exist but user try import or create new
project file will be ignored.

Next step it add functionality for recover project.

Change-Id: Ie0b1df8ad5a27109d90a3b17f56eebdafb743e2b